+++ CHANGELOG.md
@@ -31,6 +31,7 @@ The init command now also generate the configuration for tendermint, the flags `
 - [5011](https://github.com/vegaprotocol/vega/issues/5011) - Import the `protos` repo in the vega repo
 - [5774](https://github.com/vegaprotocol/vega/issues/5774) - Use `generics` for `ID` types
 - [5785](https://github.com/vegaprotocol/vega/issues/5785) - Add support form `ERC20` bridge stopped and resumed events
+- [5712](https://github.com/vegaprotocol/vega/issues/5712) - Configurable `graphql` endpoint
 
 - [5685](https://github.com/vegaprotocol/vega/issues/5685) - Migrated `apis` from `datanode v1` to `datanode v2`
 ### 🐛 Fixes